#188 - move tweets tests case
