+++ tests/extractors/content.py
@@ -424,15 +424,6 @@ class TestExtractionsRaw(TestExtractions):
         return article
 
 
-class TestArticleTweet(TestExtractionBase):
-
-    def test_tweet(self):
-        article = self.getArticle()
-        number_tweets = len(article.tweets)
-        expected_number_tweets = self.data['expected']['tweets']
-        self.assertEqual(number_tweets, expected_number_tweets)
-
-
 class TestArticleLinks(TestExtractionBase):
 
     def test_links(self):
